update crud phonebook

+++ part2/phonebook/src/components/Persons.js
@@ -21,13 +21,17 @@ const Person = ({ name, number, onDelete }) => {
 function Persons({ persons, updateOrigList, updateFilteredList }) {
   const deleteHandler = (id) => {
     console.log('delete button ', id);
-    const newList = persons.filter((x) => x.id !== id);
-    phoneService.delete(id).then((data) => {
-      console.log(data, `record ${id} deleted!`);
-      updateOrigList(newList);
-      updateFilteredList(newList);
-      console.log('NEWLIST....', newList);
-    });
+    const person = persons.find((x) => x.id === id);
+
+    if (window.confirm(`Delete ${person.name}?`)) {
+      const newList = persons.filter((x) => x.id !== id);
+      phoneService.delete(id).then((data) => {
+        console.log(data, `record ${id} deleted!`);
+        updateOrigList(newList);
+        updateFilteredList(newList);
+        console.log('NEWLIST....', newList);
+      });
+    }
   };
 
   return (
